requirements.txt
dash-bootstrap-components
matplotlib
opencv-python
numpy
pandas
pykalman
PyYAML
PyQt5
scikit-learn
scipy
seaborn
sklearn
sympy
tqdm
